Connecting code with reservation management function - zadanie JPYDZR4CC
